requirements.txt
Flask==3.1.1
WeasyPrint==63.1
python-dotenv==1.0.1
